requirements.txt
fastapi
uvicorn
openai
python-dotenv
supabase
pydantic
pandas
reportlab
python-dateutil

